✈️Refactor highlight component
